refactor(playground/web): consolidate utility functions into template-chat-utils and update imports
