+++ sites/playground/web/src/components/genui-template/JsonPatchDev.vue
@@ -5,7 +5,7 @@ import { CodeEditor } from 'monaco-editor-vue3';
 import * as jsonDiffPatch from 'jsondiffpatch';
 import * as jsonPatchFormatter from 'jsondiffpatch/formatters/jsonpatch';
 import type { JsonPatchOp } from 'jsondiffpatch/formatters/jsonpatch-apply';
-import { textToJson } from './utils';
+import { textToJson } from './template-chat-utils';
 
 const props = defineProps<{
   currentSchema: string;
